refactor(fn-key): split gesture vs OS backend → os_/fn_key/
