+++ src/dicton/os_/fn_key/_linux.py
@@ -9,65 +9,9 @@ by `DoubleTapRecognizer`, not here.
 
 from __future__ import annotations
 
-import sys
 import threading
 import time
 from collections.abc import Callable
-
-# Fn keycodes seen across vendors.
-#   143 = KEY_WAKEUP  (ThinkPad kernel remap of bare Fn)
-#   464 = KEY_FN
-#   465 = KEY_FN_ESC
-#   466 = KEY_FN_F1
-FN_KEYCODES = {143, 464, 465, 466}
-
-DOUBLE_TAP_WINDOW_S = 0.3
-
-
-class DoubleTapRecognizer:
-    """Turns a stream of taps into clean double-tap gestures.
-
-    A gesture is "settled" once no new tap arrives for ``window_s`` after the
-    last one (the timer is reset on every tap). Only a settled count of
-    **exactly two** fires ``on_double_tap``; a single tap, or a burst of three
-    or more rapid taps ("mitraille"), is discarded as noise. This is what makes
-    a deliberate tap-tap meaningful while ignoring stray presses and key
-    chatter.
-    """
-
-    def __init__(
-        self,
-        on_double_tap: Callable[[], None],
-        window_s: float = DOUBLE_TAP_WINDOW_S,
-    ) -> None:
-        self._cb = on_double_tap
-        self._window = window_s
-        self._count = 0
-        self._timer: threading.Timer | None = None
-        self._lock = threading.Lock()
-
-    def feed_tap(self) -> None:
-        with self._lock:
-            self._count += 1
-            if self._timer is not None:
-                self._timer.cancel()
-            self._timer = threading.Timer(self._window, self._settle)
-            self._timer.daemon = True
-            self._timer.start()
-
-    def _settle(self) -> None:
-        with self._lock:
-            count = self._count
-            self._count = 0
-            self._timer = None
-        if count == 2:
-            self._cb()
-
-    def stop(self) -> None:
-        with self._lock:
-            if self._timer is not None:
-                self._timer.cancel()
-                self._timer = None
 
 
 def _keyboard_devices() -> list:
@@ -111,8 +55,6 @@ def capture_keycode(timeout_s: float = 5.0) -> tuple[int, str] | None:
     key pressed. Linux/evdev only; returns None on timeout or if evdev is
     unavailable. Drains pending events first so the Enter used to arm the
     prompt isn't captured, and ignores Enter/Esc."""
-    if sys.platform != "linux":
-        return None
     try:
         from evdev import categorize, ecodes
     except ImportError:
@@ -157,23 +99,21 @@ def capture_keycode(timeout_s: float = 5.0) -> tuple[int, str] | None:
     return None
 
 
-class FnKeyListener:
+class LinuxFnKeyListener:
     """Background thread that emits a bare `on_tap()` per trigger key_down.
     Gesture meaning is decided downstream by DoubleTapRecognizer."""
 
     def __init__(
         self,
         on_tap: Callable[[], None],
-        keycodes: set[int] | None = None,
+        keycodes: set[int],
     ) -> None:
         self._on_tap = on_tap
-        self._keycodes = keycodes or FN_KEYCODES
+        self._keycodes = keycodes
         self._stop = threading.Event()
         self._thread: threading.Thread | None = None
 
     def start(self) -> bool:
-        if sys.platform != "linux":
-            return False
         try:
             import evdev  # noqa: F401
         except ImportError:
