+++ src/dicton/wizard.py
@@ -16,9 +16,10 @@ from rich.prompt import Confirm, Prompt
 from rich.table import Table
 
 from . import cleanup as cleanup_mod
-from . import fn_key, stt
+from . import stt
 from .config import CLEANUP_MODELS, ChunkParams, Config
 from .os_ import autostart as platform_mod
+from .os_ import fn_key
 from .stt import GROQ_BASE, pcm16_to_wav
 
 console = Console()
